RomoValadez_JonathanJoshua_Docker/requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
fastapi==0.79.0
h11==0.13.0
idna==3.3
psycopg2==2.9.3
pydantic==1.9.1
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
fastapi-login==1.8.3
python-multipart==0.0.5
peewee==3.15.4
pillow 